Consolidate parser and interpreter into single `cel` crate
# Conflicts:
#	antlr/Cargo.toml
#	example/Cargo.toml
