--- antlr/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "cel-parser"
-description = "An antlr-based parser for the Common Expression Language (CEL)"
-repository = "https://github.com/cel-rust/cel-rust"
-version = "0.10.1"
-authors = ["Alex Snaps <alex@wcgw.dev>"]
-edition = "2021"
-license = "MIT"
-categories = ["parsing"]
-
-[dependencies]
-antlr4rust = "0.3.0-beta3"
-lazy_static = "1.5.0"
