+++ cel/Cargo.toml
@@ -1,18 +1,17 @@
 [package]
-name = "cel-interpreter"
-description = "An interpreter for the Common Expression Language (CEL)"
+name = "cel"
+description = "A parser and interpreter for the Common Expression Language (CEL)"
 repository = "https://github.com/cel-rust/cel-rust"
 version = "0.10.0"
-authors = ["Tom Forbes <tom@tomforb.es>", "Clark McCauley <me@clarkmccauley.com>", "Alex Snaps <alex@wcgw.dev>"]
+authors = ["Clark McCauley <me@clarkmccauley.com>", "Alex Snaps <alex@wcgw.dev>"]
 edition = "2021"
 license = "MIT"
 categories = ["compilers"]
 
 [dependencies]
-cel-parser = { path = "../antlr", version = "0.10.1" }
-
+antlr4rust = "0.3.0-beta3"
+lazy_static = "1.5.0"
 nom = "7.1.3"
-
 chrono = { version = "0.4", default-features = false, features = ["alloc", "serde"], optional = true }
 regex = { version = "1.10.5", optional = true }
 serde = "1.0"
