+++ cel/src/resolvers.rs
@@ -1,5 +1,5 @@
+use crate::parser::Expression;
 use crate::{ExecutionError, FunctionContext, ResolveResult, Value};
-use cel_parser::Expression;
 
 /// Resolver knows how to resolve a [`Value`] from a [`FunctionContext`].
 /// At their core, resolvers are responsible for taking Expressions and
